G Suite: Move the isGoogleWorkspaceExtraLicence() selector to its own file
