+++ packages/calypso-products/src/product-values.ts
@@ -35,7 +35,8 @@ export { isFreeJetpackPlan } from './is-free-jetpack-plan';
 export { isFreePlanProduct } from './is-free-plan';
 export { isFlexiblePlanProduct } from './is-flexible-plan';
 export { isFreeWordPressComDomain } from './is-free-wordpress-com-domain';
-export { isGoogleWorkspace, isGoogleWorkspaceExtraLicence, isGSuite } from './is-gsuite';
+export { isGoogleWorkspace, isGSuite } from './is-gsuite';
+export { isGoogleWorkspaceExtraLicence } from './is-google-workspace-extra-license';
 export { isGoogleWorkspaceMonthly } from './is-google-workspace-monthly';
 export { isGSuiteOrExtraLicense } from './is-gsuite-or-extra-license';
 export { isGSuiteOrExtraLicenseOrGoogleWorkspace } from './is-gsuite-or-extra-license-or-google-workspace';
